# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.StockAdjustmentCreateDTO:
    properties:
      adj_adjustment_date:
        type: string
      adj_adjustment_reason:
        type: string
      adj_adjustment_type:
        type: string
      adj_approved_by:
        type: string
      adj_batch_number:
        type: string
      adj_expiry_date:
        type: string
      adj_facility_code:
        type: string
      adj_notes:
        type: string
      adj_pharmacy_id:
        description: 'Optional: for pharmacy-level adjustments'
        type: integer
      adj_product_code:
        type: string
      adj_quantity:
        type: integer
      adj_reference_number:
        type: string
      adj_system_code:
        type: string
    required:
    - adj_adjustment_date
    - adj_adjustment_reason
    - adj_adjustment_type
    - adj_batch_number
    - adj_expiry_date
    - adj_facility_code
    - adj_product_code
    - adj_quantity
    - adj_system_code
    type: object
  dto.StockAdjustmentResponseDTO:
    properties:
      adj_adjustment_date:
        type: string
      adj_adjustment_reason:
        type: string
      adj_adjustment_type:
        type: string
      adj_approved_by:
        type: string
      adj_batch_number:
        type: string
      adj_expiry_date:
        type: string
      adj_facility_code:
        type: string
      adj_notes:
        type: string
      adj_pharmacy_code:
        type: string
      adj_pharmacy_id:
        type: integer
      adj_pharmacy_name:
        type: string
      adj_product_code:
        type: string
      adj_quantity:
        type: integer
      adj_reference_number:
        type: string
      adj_system_code:
        type: string
      adj_timestamp:
        type: string
      id:
        type: integer
      sync_status:
        type: integer
      validation_status:
        type: integer
    type: object
  dto.StockAdjustmentUpdateDTO:
    properties:
      adj_quantity:
        type: integer
    type: object
  dto.StockDispensedCreateDTO:
    properties:
      dsp_batch_number:
        type: string
      dsp_dispense_date:
        type: string
      dsp_dispensed_quantity:
        type: number
      dsp_expiry_date:
        type: string
      dsp_facility_code:
        type: string
      dsp_patient_hash:
        type: string
      dsp_product_code:
        type: string
      dsp_system_code:
        type: string
    required:
    - dsp_batch_number
    - dsp_dispense_date
    - dsp_dispensed_quantity
    - dsp_expiry_date
    - dsp_facility_code
    - dsp_patient_hash
    - dsp_product_code
    - dsp_system_code
    type: object
  dto.StockDispensedResponseDTO:
    properties:
      dsp_batch_number:
        type: string
      dsp_dispense_date:
        type: string
      dsp_dispensed_quantity:
        type: number
      dsp_expiry_date:
        type: string
      dsp_facility_code:
        type: string
      dsp_patient_hash:
        type: string
      dsp_product_code:
        type: string
      dsp_system_code:
        type: string
      dsp_timestamp:
        type: string
      id:
        type: integer
      sync_status:
        type: integer
      validation_status:
        type: integer
    type: object
  dto.StockDispensedUpdateDTO:
    properties:
      dsp_dispensed_quantity:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: ict@moh.go.ug
    name: DHI
  description: Backend API for Supply Chain Repository / Bridge
  termsOfService: https://health.go.ug
  title: MoH Supply Chian Repository API
  version: "1.0"
paths:
...
  /stock/adjustment:
    get:
      parameters:
      - description: Filter by facility code
        in: query
        name: facility_code
        type: string
      - description: Filter by pharmacy ID
        in: query
        name: pharmacy_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.StockAdjustmentResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List stock adjustments
      tags:
      - StockAdjustment
    post:
      consumes:
      - application/json
      parameters:
      - description: Stock adjustment payloads
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.StockAdjustmentCreateDTO'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.StockAdjustmentResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create stock adjustment
      tags:
      - StockAdjustment
  /stock/adjustment/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete stock adjustment
      tags:
      - StockAdjustment
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StockAdjustmentResponseDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get stock adjustment by id
      tags:
      - StockAdjustment
    put:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.StockAdjustmentUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StockAdjustmentResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update stock adjustment
      tags:
      - StockAdjustment
  /stock/dispensed:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.StockDispensedResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List stock dispensed
      tags:
      - StockDispensed
    post:
      consumes:
      - application/json
      parameters:
      - description: Stock dispensed payloads
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.StockDispensedCreateDTO'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.StockDispensedResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create stock dispensed
      tags:
      - StockDispensed
  /stock/dispensed/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete stock dispensed
      tags:
      - StockDispensed
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StockDispensedResponseDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get stock dispensed by id
      tags:
      - StockDispensed
    put:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.StockDispensedUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StockDispensedResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update stock dispensed
      tags:
      - StockDispensed
...
swagger: "2.0"